[simulation]fix Jira SWBUG-8992 Scipio - Taille de l'export pour prepa

+++ src/libraries/simulation_kernel/Entities/Specialisations/LOG/LogisticLink.cpp
@@ -97,7 +97,7 @@ void LogisticLink::ReadQuota( xml::xistream& xis )
     if( quotas_.find( pDotationCategory ) != quotas_.end() )
         xis.error( "Quota already defined" );
 
-    unsigned int quantity;
+    double quantity;
     xis >> xml::attribute( "quantity", quantity );
     if( quantity < 0 )
         xis.error( "quantity is not greater or equal to 0" );
